Implement downloading feature
Add route download_file that checks, if the given user is allowed to
download a given file and responds with this file.
Add function get_file() to Database to get a file/a files owner from a
given file-id.
Add the tohex filter to the tera templating engine, that allows
displaying IDs as hex values.