Implement automatic removal of expired sessions
Add field date to struct UserSession and session-tree.
Add new tree user_session_tree to get the session_ids for a given
user_id.
Add check for age of session when getting an existing session for a new
request.
Add code to remove all expired sessions for a given user when he's
logging in.