Refactor database module
Add module fs_db to module database, which contains the struct
FsDatabase together with all functions to manipulate the saved file and
directory data in the database.
Change functions in module database to call the new functions in module
fs_db.
Replace file-tree/dir-tree in struct Database with field of type
FsDatabase.