Add routes to add and get Users
Add module controller::user. This modules includes the function
get_user_info(), which can be used to get the information of a User from
the DB, and the function add_user(), which can be used to add a new user
to the DB. These functions also include the checks for permissions.

Add method insert_user() to Database. This function is used by
controller::user::add_user().

Implemented the routes get_user_info() and add_user() in
webapi::rest_api::user_api.