Add group list/permission bits to User/Files, Dirs
Add a field "group_ids" to struct User hold references to all groups,
which the user is a member of. Adjust multiple functions to initialize
the new field and read/write this field from the DB.

Add a field permission_bits to struct Dir and File, which stores the
read and write permissions of the owning user, the owning group and all
other users for the given file/dir. To handle changes to this attribute
over the webapi the struct PermissionMsg is added. DirMsg and FileMsg
now have an attribute of this type. Multiple functions are added and
changed to handle this new attribute when a request for updating/adding
a file or dir arrives.