Replace permission bits with list of groups
Replace the system of read and write permission for owning user and
group of a FsNode with a new system, which adds a list of readable and a
list of writable groups for every FsNode.

Replace field permission_bits with Vec readable_groups and Vec
writeable_groups for all FsNode structs (File and Dir).
Change DB scheme to save the lists in a seperat table and remove the old
field from the existing tables. Adjust functions in database::fs_db to
work with the new scheme.
Change functions in module controller to work with the adjusted structs
and permission system.
Remove PermissionMsg from module webapi.