Update template and style of dirview
Replace dirview.html.tera and the style.css. The new content introduces
a new layout and design.
Update webapi::content_routes::content_pages::dir_page() to create the
necessary context for the new template. For this purpose the structs
FileContext and DirContext and the function
controller::user::resolve_user_name() are added.