Replace u64 with Id in *Msg structs
Change the type of multiple fields in the structs GroupMsg, UserMsg,
FileMsg and DirMsg from u64 to Id, Option<u64> to Option<Id> or
Option<Vec<u64>> to Option<Vec<Id>>. This is necessary to represent the
IDs as strings in the JSON representation of these structs, which allows
for a better handling of the IDs in the JS frontend, which does not
support 64 bit integers. To enable this string representation in JSON,
Serialize and Deserialize is implemented for Id.

Change the return type of all routes, that returned JSON representations
of File, Dir, Group or User to the according *Msg type. Therefore From
is implemented for all the *Msg structs to create such a struct from the
corresponding struct from model::. Addidionally Serialize is implemented
for all *Msg structs and for ID.