Fix bugs with missing files in dirview
Fix bug when adding new files: Previously a 0 was inserted as the new
files ID into the list of the parent directories childs. Now the correct
new ID is inserted.

Add size attribute to FileMsg and FileContext and change dirview
template to render a placeholder size, while we haven't implemented a
real size attribute for files.