Add Content-Type header to downloaded files
Add media_type attribute to File. Change DB scheme to save the
media_type of a File and functions in module db::fs_db to work with the
new scheme.
Update function rest_api::update_file_content to store the content type
of the pushed file. Update function rest_api::get_file_content to
responde with the correct stored Content-Type. We mostly implement this
functionylity in the corresponding functions in module controller.